feat: initial oban implementation
