+++ lib/atlas/workers/import_students_by_course.ex
@@ -1,0 +1,14 @@
+defmodule Atlas.Workers.ImportStudentsByCourse do
+  @moduledoc """
+  Worker to import students by course.
+  """
+  use Oban.Worker, queue: :import
+
+  alias Atlas.Importers.StudentsByCourse
+
+  @impl Oban.Worker
+  def perform(%Oban.Job{args: %{"file_path" => file_path}}) do
+    StudentsByCourse.import_from_excel_file(file_path)
+    :ok
+  end
+end
