+++ mix.exs
@@ -55,6 +55,7 @@ defmodule Atlas.MixProject do
       # tools
       {:credo, "~> 1.6", only: [:dev, :test], runtime: false},
       {:xlsx_reader, "~> 0.8.8"},
+      {:igniter, "~> 0.5", only: [:dev]},
 
       # monitoring
       {:telemetry_metrics, "~> 1.0"},
@@ -64,6 +65,9 @@ defmodule Atlas.MixProject do
       # server
       {:dns_cluster, "~> 0.1.1"},
       {:bandit, "~> 1.2"},
+
+      # jobs
+      {:oban, "~> 2.19"},
 
       # utilities
       {:remote_ip, "~> 1.2"},
